+++ src/CSSTransition.js
@@ -1,20 +1,24 @@
 import * as PropTypes from 'prop-types';
-import addClass from 'dom-helpers/class/addClass';
-import removeClass from 'dom-helpers/class/removeClass';
+import addOneClass from 'dom-helpers/class/addClass';
+
+import removeOneClass from 'dom-helpers/class/removeClass';
 import React from 'react';
 
 import Transition from './Transition';
 import { classNamesShape } from './utils/PropTypes';
 
+const addClass = (node, classes) => node && classes && classes.split(' ').forEach(c => addOneClass(node, c));
+const removeClass = (node, classes) => node && classes && classes.split(' ').forEach(c => removeOneClass(node, c));
+
 const propTypes = {
   ...Transition.propTypes,
 
   /**
-   * The animation classNames applied to the component as it enters or exits.
+   * The animation classNames applied to the component as it enters, exits or has finished the transition.
    * A single name can be provided and it will be suffixed for each stage: e.g.
    *
-   * `classNames="fade"` applies `fade-enter`, `fade-enter-active`,
-   * `fade-exit`, `fade-exit-active`, `fade-appear`, and `fade-appear-active`.
+   * `classNames="fade"` applies `fade-enter`, `fade-enter-active`, `fade-enter-done`,
+   * `fade-exit`, `fade-exit-active`, `fade-exit-done`, `fade-appear`, and `fade-appear-active`.
    * Each individual classNames can also be specified independently like:
    *
    * ```js
@@ -23,18 +27,22 @@ const propTypes = {
    *  appearActive: 'my-active-appear',
    *  enter: 'my-enter',
    *  enterActive: 'my-active-enter',
+   *  enterDone: 'my-done-enter,
    *  exit: 'my-exit',
    *  exitActive: 'my-active-exit',
+   *  exitDone: 'my-done-exit,
    * }}
    * ```
    *
-   * @type {{
+   * @type {string | {
    *  appear?: string,
    *  appearActive?: string,
    *  enter?: string,
    *  enterActive?: string,
+   *  enterDone?: string,
    *  exit?: string,
    *  exitActive?: string,
+   *  exitDone?: string,
    * }}
    */
   classNames: classNamesShape,
@@ -57,7 +65,7 @@ const propTypes = {
 
   /**
    * A `<Transition>` callback fired immediately after the 'enter' or
-   * 'appear' classes are **removed** from the DOM node.
+   * 'appear' classes are **removed** and the `done` class is added to the DOM node.
    *
    * @type Function(node: HtmlElement, isAppearing: bool)
    */
@@ -65,7 +73,7 @@ const propTypes = {
 
 
   /**
-   * A `<Transition>` callback fired immediately after the 'exit'  class is
+   * A `<Transition>` callback fired immediately after the 'exit' class is
    * applied.
    *
    * @type Function(node: HtmlElement)
@@ -73,8 +81,7 @@ const propTypes = {
   onExit: PropTypes.func,
 
   /**
-   * A `<Transition>` callback fired immediately after the 'exit-active' is
-   * class is applied.
+   * A `<Transition>` callback fired immediately after the 'exit-active' is applied.
    *
    * @type Function(node: HtmlElement
    */
@@ -82,7 +89,7 @@ const propTypes = {
 
   /**
    * A `<Transition>` callback fired immediately after the 'exit' classes
-   * are **removed** from the DOM node.
+   * are **removed** and the `exit-done` class is added to the DOM node.
    *
    * @type Function(node: HtmlElement)
    */
@@ -91,69 +98,18 @@ const propTypes = {
 
 /**
  * A `Transition` component using CSS transitions and animations.
- * It's inspired by the excellent [ng-animate](http://www.nganimate.org/) libary.
+ * It's inspired by the excellent [ng-animate](http://www.nganimate.org/) library.
  *
  * `CSSTransition` applies a pair of class names during the `appear`, `enter`,
  * and `exit` stages of the transition. The first class is applied and then a
- * second "active" class in order to activate the css animation.
+ * second "active" class in order to activate the css animation. After the animation,
+ * matching `done` class names are applied to persist the animation state.
  *
  * When the `in` prop is toggled to `true` the Component will get
  * the `example-enter` CSS class and the `example-enter-active` CSS class
  * added in the next tick. This is a convention based on the `classNames` prop.
  *
- * ```js
- * import CSSTransition from 'react-transition-group/CSSTransition';
- *
- * const Fade = ({ children, ...props }) => (
- *  <CSSTransition
- *    {...props}
- *    timeout={500}
- *    classNames="fade"
- *  >
- *   {children}
- *  </CSSTransition>
- * );
- *
- * class FadeInAndOut extends React.Component {
- *   constructor(...args) {
- *     super(...args);
- *     this.state= { show: false }
- *
- *     setInterval(() => {
- *       this.setState({ show: !this.state.show })
- *     }, 5000)
- *   }
- *   render() {
- *     return (
- *       <Fade in={this.state.show}>
- *         <div>Hello world</div>
- *       </Fade>
- *     )
- *   }
- * }
- * ```
- *
- * And the coorresponding CSS for the `<Fade>` component:
- *
- * ```css
- * .fade-enter {
- *   opacity: 0.01;
- * }
- *
- * .fade-enter.fade-enter-active {
- *   opacity: 1;
- *   transition: opacity 500ms ease-in;
- * }
- *
- * .fade-exit {
- *   opacity: 1;
- * }
- *
- * .fade-exit.fade-exit-active {
- *   opacity: 0.01;
- *   transition: opacity 300ms ease-in;
- * }
- * ```
+ * <iframe src="https://codesandbox.io/embed/kw8z6pp9zo?autoresize=1&fontsize=12&hidenavigation=1&moduleview=1" style="width:100%; height:500px; border:0; border-radius: 4px; overflow:hidden;" sandbox="allow-modals allow-forms allow-popups allow-scripts allow-same-origin"></iframe>
  */
 class CSSTransition extends React.Component {
   onEnter = (node, appearing) => {
@@ -180,7 +136,10 @@ class CSSTransition extends React.Component {
   }
 
   onEntered = (node, appearing) => {
+    const { doneClassName } = this.getClassNames('enter');
+
     this.removeClasses(node, appearing ? 'appear' : 'enter');
+    addClass(node, doneClassName);
 
     if (this.props.onEntered) {
       this.props.onEntered(node, appearing);
@@ -210,7 +169,10 @@ class CSSTransition extends React.Component {
   }
 
   onExited = (node) => {
+    const { doneClassName } = this.getClassNames('exit');
+
     this.removeClasses(node, 'exit');
+    addClass(node, doneClassName);
 
     if (this.props.onExited) {
       this.props.onExited(node)
@@ -218,7 +180,7 @@ class CSSTransition extends React.Component {
   }
 
   getClassNames = (type) => {
-    const { classNames } = this.props
+    const { classNames } = this.props;
 
     let className = typeof classNames !== 'string' ?
       classNames[type] : classNames + '-' + type;
@@ -226,20 +188,28 @@ class CSSTransition extends React.Component {
     let activeClassName = typeof classNames !== 'string' ?
       classNames[type + 'Active'] : className + '-active';
 
-    return { className, activeClassName }
+    let doneClassName = typeof classNames !== 'string' ?
+      classNames[type + 'Done'] : className + '-done';
+
+    return {
+      className,
+      activeClassName,
+      doneClassName
+    };
   }
 
   removeClasses(node, type) {
-    const { className, activeClassName } = this.getClassNames(type)
+    const { className, activeClassName, doneClassName } = this.getClassNames(type)
     className && removeClass(node, className);
     activeClassName && removeClass(node, activeClassName);
+    doneClassName && removeClass(node, doneClassName);
   }
 
   reflowAndAddClass(node, className) {
     // This is for to force a repaint,
     // which is necessary in order to transition styles when adding a class name.
     /* eslint-disable no-unused-expressions */
-    node.scrollTop;
+    node && node.scrollTop;
     /* eslint-enable no-unused-expressions */
     addClass(node, className);
   }